DSL: rename *Integer types to *Int
TN: R904-024

+++ langkit/expressions/boolean.py
@@ -276,7 +276,7 @@ class OrderingTest(AbstractExpression):
             lhs.type.is_big_integer_type or
             lhs.type.is_ast_node,
             'Comparisons only work on {}, {} or nodes not {}'
-            .format(T.Integer.dsl_name, T.BigInteger.dsl_name,
+            .format(T.Int.dsl_name, T.BigInt.dsl_name,
                     lhs.type.dsl_name)
         )
